pyproject.toml
[tool.poetry]
name = "ytbot"
version = "0.1.0"
description = ""
authors = ["fontanads <fontanads@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"


[tool.poetry.group.analytics.dependencies]
pandas = "<2.0"
numpy = "^1.24.2"
altair = "<5.0"


[tool.poetry.group.cloud.dependencies]
google-api-python-client = "^2.84.0"
google-auth-oauthlib = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-auth = "^2.17.2"
google-cloud-bigquery = "^3.9.0"
db-dtypes = "^1.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"
ipywidgets = "^8.0.6"
jupyterlab = "^3.6.3"
black = "^23.3.0"
flake8 = "^6.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.0"
tox = "^4.4.11"
coverage = "^7.2.3"


[tool.poetry.group.dataviz.dependencies]
streamlit = "^1.21.0"

[tool.pytest.ini_options]
addopts = "-v -s --tb=short"
testpaths = ["tests"]

# settings for black formatter
[tool.black]
extend-exclude = '''
^( # Exclude all non-Python files
   (?<!\.py)\$
)
'''
check = true
source = ["./"]
verbose = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
